+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - stacktrace of caught exceptions to event reporting (including logger)
 - ability of launching database loaders under library consumers' own supervisors
+- `wait_for_loaders/2` method for waiting on multiple databases concurrently
 ### Changed
 - log level of database loading failures from warning to error
 - HTTP loader and filesystem loader into a single loader codebase